+++ backend/leo_ems/planner/__init__.py
@@ -3,9 +3,11 @@ from .rules import ChargingRule, naechste_abfahrt, plane_garantieladung
 from .charge_control import VOLT, ChargeController, ChargeState, ChargeCommand
 from .heatpump import HeatPumpCommand, HeatPumpController
 from .batt_limit import EntladeLimitRegler, LimitEntscheid
+from .prioritaet import verbraucher_reihenfolge
 
 __all__ = [
     "berechne_ueberschuss",
+    "verbraucher_reihenfolge",
     "EntladeLimitRegler",
     "LimitEntscheid",
     "ChargingRule",